all: Initial port for gtk4
Add meson changes to build a gtk4 variant of libvte, plus
a gtk4 variant of the test application.

Make the minimal code changes required to successfully build
and run on gtk4.

No event handlers yet, and no public API to replace the gtk3
specific APIs.  That will come later.

This should be good enough to get non-terminal users of libvte
started on porting to gtk4.
terminal consumers of libvte

https://gitlab.gnome.org/GNOME/vte/-/issues/12